CONTAINERS/API/requirements.txt
fastapi
gunicorn
uvicorn[standard]
python-dotenv
mysql-connector-python

